Move comments feed to separate component
